move base64 encoding package to subfolder
