+++ encoding/base64/encoder.go
@@ -1,6 +1,8 @@
-package encoding
+package base64
 
-import "encoding/base64"
+import (
+	"encoding/base64"
+)
 
 type base64Encoder struct {
 }
@@ -13,7 +15,7 @@ func (e *base64Encoder) Decode(d string) ([]byte, error) {
 	return base64.StdEncoding.DecodeString(d)
 }
 
-func NewBase64Encoder() *base64Encoder {
+func NewEncoder() *base64Encoder {
 	e := &base64Encoder{}
 	return e
 }
