--- encoding/encoder_test.go
@@ -1,13 +1,0 @@
-package encoding_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestEncoding(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Encoding Package")
-}
